+++ src/main/webapp/taskRow.js
@@ -1,3 +1,5 @@
+const shouldDisplayRemoveTask = false;
+
 class TaskRow {
     constructor(trElement, data, rowId) {
         this.trElement = trElement;
@@ -35,30 +37,36 @@ class TaskRow {
     updateActionColumn(column, data) {
         column.innerHTML = '';
         const inputPreview = document.createElement('span');
-        inputPreview.innerText = '[I]';
+        inputPreview.className = 'badge badge-pill badge-primary';
+        inputPreview.innerText = 'Input';
         inputPreview.addEventListener('mouseover', event => showImageTooltip(event, './task/input?id=' + data.id));
         inputPreview.addEventListener('mouseout', hideImageTooltip);
         inputPreview.addEventListener('click', () => window.open('./task/input?id=' + data.id));
         column.appendChild(inputPreview);
         if (data.status === 'FINISHED') {
             const resultPreview = document.createElement('span');
-            resultPreview.innerText = '[R]';
+            resultPreview.className = 'badge badge-pill badge-success';
+            resultPreview.innerText = 'Result';
             resultPreview.addEventListener('mouseover', event => showImageTooltip(event, './task/result?id=' + data.id));
             resultPreview.addEventListener('mouseout', hideImageTooltip);
             resultPreview.addEventListener('click', () => window.open('./task/result?id=' + data.id));
             column.appendChild(resultPreview);
             const diffPreview = document.createElement('span');
             if (data.hasDiff) {
-                diffPreview.innerText = '[D]';
+                diffPreview.className = 'badge badge-pill badge-secondary';
+                diffPreview.innerText = 'Diff';
                 diffPreview.addEventListener('mouseover', event => showImageTooltip(event, './task/diff?id=' + data.id));
                 diffPreview.addEventListener('mouseout', hideImageTooltip);
                 diffPreview.addEventListener('click', () => window.open('./task/diff?id=' + data.id));
                 column.appendChild(diffPreview);
             }
-            const deleteInput = document.createElement('span');
-            deleteInput.innerText = '[X]';
-            deleteInput.addEventListener('click', () => deleteTask(data.id));
-            column.appendChild(deleteInput);
+            if (shouldDisplayRemoveTask) {
+                const deleteInput = document.createElement('span');
+                deleteInput.className = 'badge badge-pill badge-danger';
+                deleteInput.innerText = 'Remove';
+                deleteInput.addEventListener('click', () => deleteTask(data.id));
+                column.appendChild(deleteInput);
+            }
         }
     }
